requirements.txt
quixstreams
tqdm
python-dotenv
